+++ CHANGELOG.md
@@ -26,3 +26,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vector3Event class, a scriptable object event that takes a Vector3 as payload.
 - Vector3EventListener class, a component that listens for scriptable object events with Vector3 payloads.
 - IVariable interface.
+- Variable abstract class as base for all scriptable object variables.